# Implement high-performance identifiable key scanning in managed code