--- a/BuildAndTest.cmd
+++ b/BuildAndTest.cmd
@@ -4,10 +4,6 @@ setlocal
 time /t
 set RESULT=1
 
-call "%~dp0BuildAndTestRust.cmd" %*
-set RESULT=%ERRORLEVEL%
-if %RESULT% neq 0 goto :exit
-
 call powershell -ExecutionPolicy RemoteSigned -File "%~dp0scripts\BuildAndTest.ps1" %*
 set RESULT=%ERRORLEVEL%
 if %RESULT% neq 0 goto :exit