--- a/scripts/BuildAndTest.ps1
+++ b/scripts/BuildAndTest.ps1
@@ -74,29 +74,31 @@ if (-not $NoTest) {
 }
 
 Write-Information "Exporting rules data.."
-Invoke-Expression ("$RepoRoot\bld\bin\AnyCPU_Release\Microsoft.Security.Utilities.Cli\net8.0\Microsoft.Security.Utilities.Cli.exe export --output $RepoRoot\GeneratedRegexPatterns\")
+Invoke-Expression ("$RepoRoot\bld\bin\AnyCPU_$Configuration\Microsoft.Security.Utilities.Cli\net8.0\Microsoft.Security.Utilities.Cli.exe export --output $RepoRoot\GeneratedRegexPatterns\")
 if ($LASTEXITCODE -ne 0) {
     Exit-WithFailureMessage $ScriptName "Rules export failed."
 }
 
 Write-Information "Creating packages.."
-cmd.exe /c 'BuildPackages.cmd'
+cmd.exe /c "BuildPackages.cmd -Configuration $Configuration"
 if ($LASTEXITCODE -ne 0) {
     Exit-WithFailureMessage $ScriptName "Package build failed."
 }
 
-Write-Information "Building SecurityUtilitiesPackageReference.sln (dotnet)..."
-dotnet build $RepoRoot\src\SecurityUtilitiesPackageReference\SecurityUtilitiesPackageReference.sln -c $Configuration
-if ($LASTEXITCODE -ne 0) {
-    Exit-WithFailureMessage $ScriptName "Build of SecurityUtilitiesPackageReference failed."
-}
+if ($Configuration -eq 'Release') {
+    Write-Information "Building SecurityUtilitiesPackageReference.sln (dotnet)..."
+    dotnet build $RepoRoot\src\SecurityUtilitiesPackageReference\SecurityUtilitiesPackageReference.sln -c $Configuration
+    if ($LASTEXITCODE -ne 0) {
+        Exit-WithFailureMessage $ScriptName "Build of SecurityUtilitiesPackageReference failed."
+    }
 
-if (-not $NoTest) {
-    foreach ($fx in "net8.0", "net461") {
-        Write-Information "Running API examples using compiled Microsoft.Security.Utilities.Core package on $fx..."
-        Invoke-Expression "$RepoRoot\src\SecurityUtilitiesPackageReference\SecurityUtilitiesApiUtilizationExample\bin\$Configuration\$fx\SecurityUtilitiesApiUtilizationExample.exe"
-        if ($LASTEXITCODE -ne 0) {
-            Exit-WithFailureMessage $ScriptName "Microsoft.Security.Utilities.Core API example execution failed."
+    if (-not $NoTest) {
+        foreach ($fx in "net8.0", "net9.0", "net472") {
+            Write-Information "Running API examples using compiled Microsoft.Security.Utilities.Core package on $fx..."
+            Invoke-Expression "$RepoRoot\src\SecurityUtilitiesPackageReference\SecurityUtilitiesApiUtilizationExample\bin\$Configuration\$fx\SecurityUtilitiesApiUtilizationExample.exe"
+            if ($LASTEXITCODE -ne 0) {
+                Exit-WithFailureMessage $ScriptName "Microsoft.Security.Utilities.Core API example execution failed."
+            }
         }
     }
 }