--- a/src/Directory.Packages.props
+++ b/src/Directory.Packages.props
@@ -8,7 +8,6 @@
     <PackageVersion Include="Microsoft.IdentityModel.Tokens" Version="6.15.0" />
     <PackageVersion Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageVersion Include="RE2.Managed" Version="4.5.7" />
-    <PackageVersion Include="StyleCop.Analyzers" Version="1.1.118" />
     <PackageVersion Include="System.Buffers" Version="4.5.1" />
     <PackageVersion Include="System.Text.Json" Version="8.0.5" />
   </ItemGroup>