--- a/src/Microsoft.Security.Utilities.Benchmarks/SecretMaskerDetectionBenchmarks.cs
+++ b/src/Microsoft.Security.Utilities.Benchmarks/SecretMaskerDetectionBenchmarks.cs
@@ -7,22 +7,65 @@ namespace Microsoft.Security.Utilities.Benchmarks
 {
     public abstract class SecretMaskerDetectionBenchmarks
     {
-        // The # of iterations of the scan to run.
-        protected virtual int Iterations => 10;
-
         // The size of randomized data to add as a prefix
         // for every secret. This is intended to make positive
         // hit less concentrated in the profiling.
-        protected virtual int SecretPrefixSize => 100 * 1024;
+        private const int SecretPrefixSize = 100 * 1024;
 
-        protected virtual string ScanContentPrefix => GenerateRandomData(SecretPrefixSize);
+        private readonly List<string> _randomExamples;
+        private readonly List<string> _nonRandomExamples;
 
-        private static string GenerateRandomData(int size)
+        // Use a fixed seed to ensure different runs use the same data.
+        private readonly Random rng = new Random(Seed: 42);
+
+        protected SecretMaskerDetectionBenchmarks()
         {
-            var random = new Random();
-            var data = new byte[size];
-            random.NextBytes(data);
-            return Convert.ToBase64String(data);
+            _randomExamples = GenerateExamples(random: true);
+            _nonRandomExamples = GenerateExamples(random: false);
+        }
+
+        private string GeneratePrefix(int size, bool random)
+        {
+            string prefix;
+
+            if (random)
+            {
+                // This is close to the worst case scenario. Every base64
+                // alphabet char is equally likely to appear in the input,
+                // making signature sniffing harder.
+                var data = new byte[size / 4 * 3]; // Account for base64 encoding overhead.
+                rng.NextBytes(data);
+                prefix = Convert.ToBase64String(data);
+            }
+            else
+            {
+                // This is close to the best case scenario. This char does not appear in
+                // any identifiable key signature, making signature sniffing easier.
+                prefix = new string('%', size);
+            }
+
+            if (prefix.Length != size)
+            {
+                throw new InvalidOperationException("Something is wrong in math above.");
+            }
+
+            return prefix;
+        }
+
+        private List<string> GenerateExamples(bool random)
+        {
+            var examples = new List<string>();
+            var prefix = GeneratePrefix(SecretPrefixSize, random);
+
+            foreach (var pattern in RegexPatterns)
+            {
+                foreach (var example in pattern.GenerateTruePositiveExamples())
+                {
+                    examples.Add($"{prefix} {example}");
+                }
+            }
+
+            return examples;
         }
 
         // Whether to generate correlating ids for each match.
@@ -33,62 +76,73 @@ namespace Microsoft.Security.Utilities.Benchmarks
         protected abstract IEnumerable<RegexPattern> RegexPatterns { get; }
 
         [Benchmark]
-        public void UseIdentifiableScan()
+        public void IdentifiableScan_RandomPrefix()
         {
             var masker = new IdentifiableScan(RegexPatterns,
                                               GenerateCorrelatingIds);
 
-            ScanTestExamples(masker);
+            ScanTestExamples(masker, _randomExamples);
         }
 
+        [Benchmark]
+        public void IdentifiableScan_NonRandomPrefix()
+        {
+            var masker = new IdentifiableScan(RegexPatterns,
+                                              GenerateCorrelatingIds);
+
+            ScanTestExamples(masker, _nonRandomExamples);
+        }
 
         [Benchmark]
-        public void UseCachedDotNet()
+        public void SecretMasker_CachedDotNet_RandomPrefix()
         {
             var masker = new SecretMasker(RegexPatterns,
                                           GenerateCorrelatingIds,
                                           CachedDotNetRegex.Instance);
 
-            ScanTestExamples(masker);
+            ScanTestExamples(masker, _randomExamples);
         }
 
         [Benchmark]
-        public void UseRE2()
+        public void SecretMasker_CachedDotNet_NonRandomPrefix()
+        {
+            var masker = new SecretMasker(RegexPatterns,
+                                          GenerateCorrelatingIds,
+                                          CachedDotNetRegex.Instance);
+
+            ScanTestExamples(masker, _nonRandomExamples);
+        }
+
+        [Benchmark]
+        public void SecretMasker_RE2_RandomPrefix()
         {
             var masker = new SecretMasker(RegexPatterns,
                                           GenerateCorrelatingIds,
                                           RE2RegexEngine.Instance);
 
-            ScanTestExamples(masker);
+            ScanTestExamples(masker, _randomExamples);
         }
 
-        protected virtual void ScanTestExamples(ISecretMasker masker)
+        [Benchmark]
+        public void SecretMasker_RE2_NonRandomPrefix()
         {
-            int globalCount = 0;
+            var masker = new SecretMasker(RegexPatterns,
+                                          GenerateCorrelatingIds,
+                                          RE2RegexEngine.Instance);
 
-            for (int i = 1; i <= Iterations; i++)
+            ScanTestExamples(masker, _nonRandomExamples);
+        }
+
+        private void ScanTestExamples(ISecretMasker masker, List<string> examples)
+        {
+            foreach (string example in examples)
             {
-                int localCount = 0;
-
-                foreach (var regexPattern in RegexPatterns)
+                int count = masker.DetectSecrets(example).Count();
+                if (count != 1)
                 {
-                    foreach (string example in regexPattern.GenerateTruePositiveExamples())
-                    {
-                        localCount++;
-
-                        // Demonstrate classification/detection only.
-                        int count = masker.DetectSecrets($"{ScanContentPrefix} {example}").Count();
-
-                        if (count != 1)
-                        {
-                            throw new InvalidOperationException($"Regex {regexPattern.Name} failed to detect example {example}");
-                        }
-
-                        globalCount += count;
-                    }
+                    throw new InvalidOperationException("Failed to  detect example.");
                 }
             }
-            Console.WriteLine($"Total matches: {globalCount}");
         }
     }
 }