… .github/workflows/dotnet.yml
… GeneratedRegexPatterns/HighConfidenceSecurityModels.json
… GeneratedRegexPatterns/PreciselyClassifiedSecurityKeys.json
… docs/ReleaseHistory.md
… src/Microsoft.Security.Utilities.Core/HighPerformanceScanner.cs
… src/Microsoft.Security.Utilities.Core/IdentifiableScan.cs
… src/Microsoft.Security.Utilities.Core/Microsoft.Security.Utilities.Core.csproj
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/Azure32ByteIdentifiableKey.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/Azure64ByteIdentifiableKey.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/HighPerformancePattern.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/IHighPerformanceScannableKey.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/IdentifiableKey.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_156_AzureClientAppIdentifiableCredentials.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_158_AzureFunctionIdentifiableKey.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_166_AzureSearchIdentifiableQueryKey.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_176_AzureContainerRegistryIdentifiableKey.cs
… src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_200_CommonAnnotatedSecurityKey.cs
… src/Microsoft.Security.Utilities.Core/RegexPattern.cs
… src/Microsoft.Security.Utilities.Core/WellKnownRegexPatterns.cs
… src/Microsoft.Security.Utilities.Core/stylecop.json
… src/Microsoft.Security.Utilities.sln
… src/Tests.Microsoft.Security.Utilities.Core/CompiledHighPerformancePatternTests.cs
… src/Tests.Microsoft.Security.Utilities.Core/HighPerformanceScannerTests.cs
… src/Tests.Microsoft.Security.Utilities.Core/IdentifiableScanTests.cs
… src/Tests.Microsoft.Security.Utilities.Core/Tests.Microsoft.Security.Utilities.Core.csproj
… src/security_utilities_rust_ffi/Cargo.toml
… src/security_utilities_rust_ffi/src/lib.rs