This change removes the interop with the rust API from `IdentifiableScan` and implements similar high-performance scanning techniques in managed code. This reduces the size of the NuGet package tremendously, from ~7 MB to ~200 KB, and adds support for high-performance scanning on non x86/x64 CPUs and non-Windows operating systems.

Performance is generally better than before for managed callers of `IdentifiableScan` with significant gains in most cases. See benchmarks below. 

Code generation is used to avoid having to keep the high-performance scanning code in sync when patterns are changed. See also extensive code comments on the approach.

For now, the streaming capability of `IdentifiableScan` has been removed, but this can be re-added as the managed algorithm is actually streaming friendly by only ever regex-matching against short sub-ranges of input. This was held back for lack of time and discovery that our only partner that calls this API does not use the streaming. #155

Finally, the `StyleCop` analyzer is removed because it broke on newer C# constructs. This will be replaced with a `dotnet format` check like the one in `microsoft/cask` in a follow-up change.

# Benchmark

Scan all generated true positive examples of all identifiable patterns, with 100 KB of other data before each match.

`RandomPrefix` is random base64 data and `NonRandomPrefix` is all `&` characters. The former stresses the strength of signature sniffing algorithm which varies between .NET 8, .NET 9, .NET Framework, and Rust. The latter contains nothing that could be part of a signature except in a match, so just about any signature sniffing algorithm should perform well. We also loosely refer to these as the 'best' and 'worst' cases in notes below.

# Environment
```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.26100.3194)
13th Gen Intel Core i9-13900K, 1 CPU, 32 logical and 24 physical cores
```

# Summary of results

* .NET 9: ~21x faster on worst case, ~6.9x faster on best case.
* .NET 8: ~1.5x faster on worst case, ~8.4x faster on best case. 
* .NET Framework: ~2.4x faster on worst case, ~17x faster on best case.

😍

# Full results

## .NET 9

### Before this PR

| Method                                    | Mean        | Error    | StdDev   |
|------------------------------------------ |------------:|---------:|---------:|
| IdentifiableScan_RandomPrefix             |   189.72 ms | 1.931 ms | 1.806 ms |
| IdentifiableScan_NonRandomPrefix          |    61.35 ms | 0.999 ms | 0.935 ms |
| SecretMasker_CachedDotNet_RandomPrefix    |   264.72 ms | 5.072 ms | 4.981 ms |
| SecretMasker_CachedDotNet_NonRandomPrefix |   263.14 ms | 1.877 ms | 1.755 ms |
| SecretMasker_RE2_RandomPrefix             | 1,031.28 ms | 9.023 ms | 7.998 ms |
| SecretMasker_RE2_NonRandomPrefix          | 1,026.15 ms | 3.668 ms | 3.251 ms |


### After this PR

| Method                                    | Mean         | Error     | StdDev    |
|------------------------------------------ |-------------:|----------:|----------:|
| IdentifiableScan_RandomPrefix             |     9.186 ms | 0.1749 ms | 0.1718 ms |
| IdentifiableScan_NonRandomPrefix          |     8.843 ms | 0.0890 ms | 0.0832 ms |
| SecretMasker_CachedDotNet_RandomPrefix    |   260.554 ms | 2.6871 ms | 2.3821 ms |
| SecretMasker_CachedDotNet_NonRandomPrefix |   261.570 ms | 2.7378 ms | 2.5609 ms |
| SecretMasker_RE2_RandomPrefix             | 1,020.852 ms | 2.4914 ms | 2.2086 ms |
| SecretMasker_RE2_NonRandomPrefix          | 1,033.865 ms | 3.5205 ms | 3.2931 ms |


## .NET 8

### Before this PR

| Method                                    | Mean        | Error    | StdDev   |
|------------------------------------------ |------------:|---------:|---------:|
| IdentifiableScan_RandomPrefix             |   193.15 ms | 3.466 ms | 3.073 ms |
| IdentifiableScan_NonRandomPrefix          |    61.72 ms | 0.835 ms | 0.781 ms |
| SecretMasker_CachedDotNet_RandomPrefix    |   271.70 ms | 0.940 ms | 0.734 ms |
| SecretMasker_CachedDotNet_NonRandomPrefix |   264.04 ms | 4.581 ms | 4.285 ms |
| SecretMasker_RE2_RandomPrefix             | 1,032.38 ms | 3.171 ms | 2.811 ms |
| SecretMasker_RE2_NonRandomPrefix          | 1,020.46 ms | 2.632 ms | 2.462 ms |

### After this PR

| Method                                    | Mean         | Error     | StdDev    |
|------------------------------------------ |-------------:|----------:|----------:|
| IdentifiableScan_RandomPrefix             |   128.917 ms | 1.0331 ms | 0.9158 ms |
| IdentifiableScan_NonRandomPrefix          |     7.374 ms | 0.1434 ms | 0.2103 ms |
| SecretMasker_CachedDotNet_RandomPrefix    |   273.672 ms | 3.2205 ms | 3.0125 ms |
| SecretMasker_CachedDotNet_NonRandomPrefix |   259.952 ms | 2.7058 ms | 2.5310 ms |
| SecretMasker_RE2_RandomPrefix             | 1,025.887 ms | 4.9130 ms | 4.5956 ms |
| SecretMasker_RE2_NonRandomPrefix          | 1,018.437 ms | 3.9070 ms | 3.6546 ms |

## .NET Framework

### Before this PR

| Method                                    | Mean       | Error    | StdDev   |
|------------------------------------------ |-----------:|---------:|---------:|
| IdentifiableScan_RandomPrefix             |   706.1 ms |  4.52 ms |  4.23 ms |
| IdentifiableScan_NonRandomPrefix          |   679.0 ms |  3.65 ms |  3.42 ms |
| SecretMasker_CachedDotNet_RandomPrefix    | 4,262.5 ms | 29.07 ms | 25.77 ms |
| SecretMasker_CachedDotNet_NonRandomPrefix | 5,038.9 ms | 10.80 ms | 10.10 ms |
| SecretMasker_RE2_RandomPrefix             | 5,052.6 ms | 13.99 ms | 12.40 ms |
| SecretMasker_RE2_NonRandomPrefix          | 5,877.4 ms | 51.33 ms | 48.02 ms |


## After this PR

| Method                                    | Mean        | Error     | StdDev    |
|------------------------------------------ |------------:|----------:|----------:|
| IdentifiableScan_RandomPrefix             |   291.25 ms |  3.845 ms |  3.596 ms |
| IdentifiableScan_NonRandomPrefix          |    39.15 ms |  0.226 ms |  0.200 ms |
| SecretMasker_CachedDotNet_RandomPrefix    | 4,221.73 ms |  6.341 ms |  5.932 ms |
| SecretMasker_CachedDotNet_NonRandomPrefix | 5,012.14 ms | 42.388 ms | 39.649 ms |
| SecretMasker_RE2_RandomPrefix             | 5,077.60 ms | 39.867 ms | 33.291 ms |
| SecretMasker_RE2_NonRandomPrefix          | 5,891.66 ms | 32.403 ms | 30.309 ms |


# Notes

- `SecretMasker` is unchanged in this PR while `IdentifiableScan` that used to use rust interop before this PR and is now implemented in managed code. One of the next steps will be to eliminate these competing API and have a single entry point that is as fast as possible for the patterns its given. #154.

- We should also apply similar signature sniffing techniques to more than identifiable keys: #156.

- "Before this PR" includes a bug-fix where IdentifiableScan wasn't using rust for HISv2 keys due to 'JQQJ9' vs 'JQQJ' signature mismatch bug. See #129 that was closed as it is subsumed by this PR. The actual numbers for IdentifiableScan without this fix are much worse, but not shown above.

- The fact that .NET 8 does slightly better than .NET 9 on the best case is actually expected. In this synthetic case, multi-char search is as good as multi-string search, and of course multi-char search is faster than multi-string search. The remarkable thing about the .NET 9 implementation is that it doesn't slow down as you throw it harder cases.

- These numbers are not representative of the rust library's performance in isolation. Much of the cost measured here is due to interop overhead, and much of the overhead is due to deficiencies in our own code and not intrinsic costs of interop in the CLR. Nevertheless, it would be a lot of work to tune the interop and then we'd still have the huge nupkg and non-portability. From other experiment not show here, even if the interop overhead were reduced to zero, we'd still not be materially faster in the best cases and we'd be even still be slower for some combinations of input and .NET version.

- It would be an interesting follow-up to make the benchmark have a parameter that is the probability of encountering a leading signature char capital like 'A', 'J', 'Q', '~', '+'. Currently we test RadomPrefix=P(3/64) and NonRandomPrefix=P(0). We could also try to determine what a realistic probability is for real data getting scanned in order to have a synthetic benchmark that is somewhat representative of the average case.

- What the heck is up with RE2 being slower, even on .NET Framework? #158 

- The pattern for making the high-performance regexes from the general regexes could be better: #151