@michaelcfanning 2025-02-25
>             int iterations = 1000;

for god's sake, drop this value to 1.

---
Refers to: src/Tests.Microsoft.Security.Utilities.Core/IdentifiableScanTests.cs:50 in 7cbf96a. [](commit_id = 7cbf96a3641ca8e5208c94ff2119b7b8ce288bd8, deletion_comment = False)