@nguerrera 2025-03-03
@michaelcfanning @rwoll Please take another look. I have:

- addressed all comments
- fixed up issues with benchmark
- put benchmark results and more details in description
- added some tests

I have changed enough here that a full review pass over the entire change is warranted.