@nguerrera 2025-03-03
I will need to adjust the PR validation settings on GitHub to expect the debug and release .NET runs. I will wait until this is approved and do it right before merging. Otherwise, I will break other PRs.